ARMv7Assembler: suppress a -Wnarrowing warning when compiling with GCC https://bugs.webkit.org/show_bug.cgi?id=157576
Patch by Zan Dobersek <zdobersek@igalia.com> on 2016-05-11
Reviewed by Csaba Osztrogonác.

* assembler/ARMv7Assembler.h:
(JSC::ARMv7Assembler::revertJumpTo_movT3movtcmpT2): Explicitly cast the
`OP_CMP_reg_T2 | left` value to uint16_t, avoiding a narrowing conversion
warning that's being reported when compiling with GCC. The warning is sprung
due to RegisterID (which is the type of `left`) being an enum based on int,
even when the enum itself only declares 23 values.

Canonical link: https://commits.webkit.org/175743@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200742 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/assembler/ARMv7Assembler.h
@@ -1332,7 +1332,7 @@ public:
             twoWordOp5i6Imm4Reg4EncodedImmSecond(right, lo16),
             twoWordOp5i6Imm4Reg4EncodedImmFirst(OP_MOVT, hi16),
             twoWordOp5i6Imm4Reg4EncodedImmSecond(right, hi16),
-            OP_CMP_reg_T2 | left
+            static_cast<uint16_t>(OP_CMP_reg_T2 | left)
         };
         performJITMemcpy(address, instruction, sizeof(uint16_t) * 5);
         cacheFlush(address, sizeof(uint16_t) * 5);
